Added support for the readyau command
The new command accepts two arguments in the form of times. One to readyat and one to readyuntil. The time formats reamine the same as in other commands. Updated the checkRAL.js file to properly recognize the end of a readyuntil command as well as the end of a pending readyau command.