Array Storage and Opinion Scoring
This update fixes some file saving issues. Instead of the bot saving parties as strings it now saves them as arrays the first element being the user's name and the second being the id. this foregoes any issues that may have arisen from the users having the strings " : " or ", " in thier name allowing for a more robust method of storing information. the same goes for the partyarchive.json file now saving the partylist in an array as well. Aside from that I have also added support for the bot to have opinions of individuals based on the words they frequently use in messages. This is at the beginning of the generics\nicebot.js file. It checks each word of each meassage and decides whether or not to add a point or remove a point based on what words are in the meassage. this can later be used to offer dynamic responses based on the bot's opinion of certain people. The bot opinion is currently not used within the rest of the file and will have to be updated based on your wants and needs.