Docker Linter Prettier Pre-Commits
Laid out loose plans in the readme

we're doing docker. i set up a docker compose to run one service. that
service is the api service. it doesnt do anything real yet, but it
exists and has a hello world endpoint.

notably! the expectation is that going forward all new code in docker
services will be typescript.

i also set up a linter and prettier. those run automatically via a pre
commit hook with husky.

I also tidied up the .gitignore a tad. apparently the node_modules
folder was committed. that shouldn't have been the case. now it's not.