Postgres Dockerized, first setup steps
OK so I made a new service for postgres. it's not used yet, but it
exists.

when ran for the first time, the database initializes with a user called
readybot; username/password can be configured in the root .env

I also set up migrations. they aren't used yet, but they exist, and they
seem to work when I test them. I was going to use them to create that
readybot user, but you cant read from .env for .sql files, and dbmate
migrations dont support .sh files in any nice way. ultimately I think
using the init folder for the user is better practice anyway.

Updated the makefile with some more shortcuts - one for creating a new
migration, and one for running migrations on the currently running
docker things.

I wanna throw in something here about why i put the migrations folder
and logic in the api. i thought really hard on this, and went back and
forth a couple times. Ultimately I decided that putting it in the db
service was bad practice because it adds a lot of complexity to
something I can otherwise really really trust, and that creating a new
migrations service was overkill. and so i put it in the api service. it
feels a little weird to have the api service to ALSO be in charge of
migrations, but in a world of no perfect choices it seems to be the most
tolerable choice.

All in all, I think this is a nice foundational step for me to build
future things on top of.