backups backups backups gdrive
OK. so. basically. theres backups now. backups go to gdrive. rclone is
used to get files to gdrive

theres a new docker service. it waits for postgres to start, and then it
starts the backup script. the backup script uses rclone to send a backup
to gdrive once every 24 hours, and the 14 most recent backups are
retained in a short term folder. when there are more than 14 short term
backups, the oldest ones are either:
1) promoted to long term if they're more than ~5 months older than the most recent
   long term backup
2) pruned (read: deleted, killed, destroyed) otherwise.

updated makefile with convenient options for nuking everything. i got
tired of typing.

in case its not clear, backups are stored in `root/db-backups/<dbname>/backups`.
and then they're shipped to gdrive. backups on the local machine are
capped at the most recent 10

also I tidied up the readme a bit to be more readable. and the setup
instructions are in a pretty solid state - i had to set up my dev
environment on a new pc and everything i did is in the setup section.

i also threw together a test that tests that the backup logic works. its
just a bash file, and isnt automatically ran anywhere. but it exists
which is pretty cool