Restores and Migrations Cleanup
So migrations on the raspi are weird because theres no nice binary that
seems to work on my raspi out of the gate. and so previously, I made it
so that we literally download the dbmate source and compile it ourselves
whenevere we run migrations. Which seemed to be a lot. (on the raspi
this takes like 7 minutes). I set out here to change that/make it not
take 7 minutes every time I want to run migrations on the raspi.
However, I've realized that docker automatically does some caching
stuff, and so it will build the binary only if theres been a change in
the dockerfile or if the dbmate repo is updated. I checked the repo, and
it seems like they only update the repo maybe once or twice a month, and
so I've decided that its just fine as is. Good enough, so to speak.

Backups and restores have been tweaked a bit. Instead of passing in a
--prod flag to indicate we're doing a prod backup, we pass in a --test
flag to indicate we're doing a test. My reasoning here is that I
sometimes do backups in my dev environment, and doing --prod to do those
backups doesn't make sense.

Also, restoring a backup now downloads the backup into the backups
folder for that particular db, meaning that they'll get cleaned up with
the regular local backups. Good enough for me.

I threw in a make target for running the backup tests, because I
would hate to forget they exist and then mess around with backups and
then oopsie they're broken

I also threw in a make target to redeploy a service. because that's
nice to have.

I tore out a bunch of dead code from the shitbot, and the shitbot now
pulls all env vars from the .env rather than from the
data/config.json or whatever. because that was weird and gross and bad.