data file -> postgres
Ok so, broadly the goal right now is to try and get the shitbot into docker so
that everything is in docker. that's a nice progress marker to aim for.

Notably, the shitbot still stores some of its data in a json file.  that's not
great, because I want backups to backup everything, but if some of the critical
data is in a json file that isnt backed up then that is bad. so, while I
certainly COULD just stuff the main.js AND the data file into docker, I felt
that it would be better if I moved the data into postgres first. So this pr is
me moving that data file into postgres.

The core changes in this pr include:
- added a migration to add a table for themoji/playlist id pairs
- wrote `migratejsontopostgres.sh` which migrates the json data into postgres
- updated the shitbot to read/write data from/to the db instead of the file
- there was a field in the data file called `mostRecentOrderTime` which was set
  and never read, so I didn't move that into postgres and I removed the writes
to the file. figure the field will die with the data file.

Additionally, I did the following less related things:
- updated a bunch of comments
- added a launch.json so the vscode debugger can be used
- refactored the `sendBallots` code to be async and a little more readable
- removed some (most? all?) of the "test mode" stuff, because now that sort of
  idea is managed via separate .env's

Some final notable thoughts: I was originally planning to port the data file
into the db and then access the data from the db whenever it is needed, but that
meant we had to have a bunch of async db calls happening in a lot of not-async
places. and i could make those places async, but thats a bigger lift and i think
that's more complicated than i really want this pr to be. so what i'm doing
instead is putting the data in the db, but just reading and updating the db data
when new themes are created/destroyed. so its the same sort of cache in memory
strategy as we've had with the data file. idk. seems fine.