Unit Tests For GET /scores
Some unit test dev logistics/tools:
- When you run unit tests locally, migrations are now ran on the db
- CI/CD now has env vars for POSTGRES_USER and POSTGRES_PASSWORD, this
  is so tests can access the db directly
- Added a launch config for debugging unit tests in the vscode debugger
	- split out the make target for `make unit-tests-api` into a prep target
and a target that preps and runs tests. This is so the debugger can
call the prep target before initializing unit tests in debug. The idea here is
that this way the two ways of prepping for unit tests (debug and non-debug) stay
in sync. it would suck to update unit tests one way, and have them not work in
the other.
	- Updated a bunch of jest config shenanigans so that the
	  debugger can map ts files to the compiled js files, and vice
versa
- NOTE: i've configured tests to run on one worker so that these tests
  run sequentially. I can imagine a future where we mess with that to
get some tests running in parallel, but I didn't want to shake that boat
in this pr.

Tidied+finished unit tests for GET /scores:
- Tests for existing uris, nonexisting uris, aggregate scores within,
  across, and outside of intervals, and that scores with different themojis are in
fact aggregated separately correctly
- Tests for bad inputs
	- uri & themoji are required (400 if missing)
	- interval is optional, defaults to 618 years if not supplied,
	  and 500's if supplied interval is invalid

Tweaked the actual GET /scores endpoint:
- Updated the comment - eventually i hope to standardize the
  documentation of these endpoints, and have like swagger or something.
But for now this will do.
- We were checking for if (result.rows.length === 0), but because we're
  aggregating scores into two totals (interval and total), there will
always be exactly two rows. So i removed the check entirely.
- We weren't converting the score numbers from the string result to
  numbers, now we are.

Additionally, updated the score insertion code to use postgres's NOW()
rather than generating a now and plugging it in. this should avoid weird
timezone things. because the only one generating a time is postgres.

Finally, you may notice the commit history is a bit bungled. this is because, as
jasper-rutherford, I have supreme power in this repo, and main's branch
protection rule cannot stop me, and i accidentally pushed a bunch of changes
straight on to main. tut tut. i've since changed the settings such that I can no
longer bypass the rules. go team!